Refactor with factor pattern
