+++ packages/api/src/services/integrations/pocket.ts
@@ -1,0 +1,5 @@
+import { IntegrationService } from './integration'
+
+export class PocketIntegration extends IntegrationService {
+  name = 'POCKET'
+}
